+++ lib/src/interfaces/interfaces.dart
@@ -2,7 +2,8 @@ library interfaces;
 
 import 'dart:typed_data';
 
-import 'package:blockchain_utils/tuple/tuple.dart';
+import 'package:blockchain_utils/utils/utils.dart';
+import 'package:passkeys/types.dart';
 import 'package:web3dart/crypto.dart';
 import 'package:web3dart/web3dart.dart';