+++ lib/src/web3_signers_base.dart
@@ -5,7 +5,6 @@ import 'dart:convert';
 import 'dart:math';
 import 'dart:typed_data';
 
-import 'package:blockchain_utils/bip/mnemonic/mnemonic.dart';
 import 'package:blockchain_utils/blockchain_utils.dart';
 import 'package:passkeys/authenticator.dart';
 import 'package:passkeys/types.dart';